# [oneDNN] Make cross_entropy calculations parallel on CPU