@cantonios 2024-03-14
> Hi @cantonios, I just ran benchmarks again and it still has better performance. Can you please give me more details to reproduce the worse performance case? Was it on CPU or GPU?

CPU.  It's only the CPU kernel that's modified, and the benchmark numbers in the revert commit list the CPU benchmark numbers.