@cantonios 2024-03-21
> @cantonios One more thing, can you please share system specifications that were used to run this benchmark? What CPU did you use?

Similar results on both haswell and skylake.