@cantonios 2024-03-25
> > > @cantonios One more thing, can you please share system specifications that were used to run this benchmark? What CPU did you use?
> > 
> > 
> > Similar results on both haswell and skylake.
> 
> @cantonios I ran it on skylake machine and got better results after optimization:
> 
> Benchmark	Items with optimization	Items without optimization
> BM_Xent_16_10000_cpu_float/real_time	182.368M/s	103.502M/s
> BM_Xent_32_10000_cpu_float/real_time	289.048M/s	176.654M/s
> BM_Xent_64_10000_cpu_float/real_time	428.617M/s	298.246M/s
> BM_Xent_16_10000_cpu_bfloat16/real_time	151.684M/s	86.4855M/s
> BM_Xent_32_10000_cpu_bfloat16/real_time	358.028M/s	149.516M/s
> BM_Xent_64_10000_cpu_bfloat16/real_time	390.276M/s	254.699M/s
> I'm trying to get down to the bottom of why we're getting better performance. Can you please provide me more details about the system that was used for benchmarking and significant run options such as cores used? If possible it would be great if I could use the machine you ran it on.

It's on our Google perflab machines.  You cannot access them.  This is without oneDNN, using TF's defaults.

Plus, this is only half the battle.  We also received reports from several teams within Google that their models no longer converged after this change - so there are numerical differences as well.